move buildKubeConfig to separate class
