+++ src/main/java/org/jenkinsci/plugins/kubernetes/auth/KubernetesAuthCertificate.java
@@ -7,7 +7,7 @@ import io.fabric8.kubernetes.client.ConfigBuilder;
 import io.fabric8.kubernetes.client.internal.SerializationUtils;
 import org.jenkinsci.plugins.kubernetes.credentials.Utils;
 
-public class KubernetesAuthCertificate implements KubernetesAuth {
+public class KubernetesAuthCertificate extends KubeConfigBuilder implements KubernetesAuth {
     private final String certificate;
 
     private final String key;
@@ -26,30 +26,12 @@ public class KubernetesAuthCertificate implements KubernetesAuth {
     }
 
     @Override
-    public String buildKubeConfig(String serverUrl, String caCertificate) throws JsonProcessingException {
-        io.fabric8.kubernetes.api.model.ConfigBuilder configBuilder = new io.fabric8.kubernetes.api.model.ConfigBuilder();
-        // setup cluster
-        Cluster cluster = new Cluster();
-        cluster.setServer(serverUrl);
-        if (caCertificate != null && !caCertificate.isEmpty()) {
-            cluster.setCertificateAuthorityData(Utils.wrapCertificate(caCertificate));
-        } else {
-            cluster.setInsecureSkipTlsVerify(true);
-        }
-        configBuilder.addNewCluster().withName("k8s").withCluster(cluster).endCluster();
-
-        // setup user (class-specific)
-        AuthInfoBuilder authInfoBuilder = new AuthInfoBuilder();
+    public void decorate(AuthInfoBuilder authInfoBuilder) {
         authInfoBuilder.withClientCertificateData(getCertificate()).withClientKeyData(getKey());
-
-        configBuilder.addNewUser().withName("cluster-admin").withUser(authInfoBuilder.build()).endUser();
-        // setup context
-        configBuilder.addNewContext().withName("k8s").withNewContext().withCluster("k8s").withUser("cluster-admin").endContext().endContext();
-        return SerializationUtils.getMapper().writeValueAsString(configBuilder.build());
     }
 
     @Override
-    public ConfigBuilder decorate(ConfigBuilder builder) throws KubernetesAuthException {
+    public ConfigBuilder decorate(ConfigBuilder builder) {
         builder.withClientCertData(certificate);
         builder.withClientKeyData(key);
         if (password != null) {
